+++ packages/terra-date-input/src/DateInput.jsx
@@ -269,12 +269,18 @@ class DateInput extends React.Component {
    * @param {Object} event Event object generated from the event delegation.
    */
   handleDayKeyDown(event) {
-    let stateValue = this.state.day || '0';
+    let stateValue = this.state.day || '';
     const previousStateValue = stateValue;
     const displayFormat = DateInputUtil.computedDisplayFormat(this.props.displayFormat, this.props.intl.locale);
 
     // prevent e and . characters from being entered into number input on keyDown
-    if (event.keyCode === 69 || event.keyCode === 190) {
+    if (event.keyCode === KeyCode.KEY_E || event.keyCode === KeyCode.KEY_PERIOD) {
+      event.preventDefault();
+      return;
+    }
+
+    // prevent + and - characters from being entered into number input on keyDown
+    if (event.keyCode === DateInputUtil.keyCode.KEY_MINUS || event.keyCode === DateInputUtil.keyCode.KEY_PLUS) {
       event.preventDefault();
       return;
     }
@@ -305,12 +311,18 @@ class DateInput extends React.Component {
    * @param {Object} event Event object generated from the event delegation.
    */
   handleYearKeyDown(event) {
-    let stateValue = this.state.year || '0';
+    let stateValue = this.state.year || '';
     const previousStateValue = stateValue;
     const displayFormat = DateInputUtil.computedDisplayFormat(this.props.displayFormat, this.props.intl.locale);
 
     // prevent e and . characters from being entered into number input on keyDown
-    if (event.keyCode === 69 || event.keyCode === 190) {
+    if (event.keyCode === KeyCode.KEY_E || event.keyCode === KeyCode.KEY_PERIOD) {
+      event.preventDefault();
+      return;
+    }
+
+    // prevent + and - characters from being entered into number input on keyDown
+    if (event.keyCode === DateInputUtil.keyCode.KEY_MINUS || event.keyCode === DateInputUtil.keyCode.KEY_PLUS) {
       event.preventDefault();
       return;
     }
@@ -382,11 +394,10 @@ class DateInput extends React.Component {
     }
 
     const inputValue = event.target.value;
-    const maxValue = 9999;
 
     // When 'Predictive text' is enabled on Android the maxLength attribute on the input is ignored so we have
     // to check the length of inputValue to make sure that it is less then 4.
-    if (inputValue.length > 5 || Number(inputValue) > maxValue) {
+    if (inputValue.length > 5 || Number(inputValue) > DateInputUtil.MaxYearValue || (inputValue.length === 4 && Number(inputValue) < DateInputUtil.MinYearValue)) {
       return;
     }
 
@@ -501,7 +512,7 @@ class DateInput extends React.Component {
           disabled={this.props.disabled}
           aria-describedby={ariaDescriptionId}
         >
-          <option value="" hidden>{this.props.intl.formatMessage({ id: 'Terra.date.input.monthPlaceholder' })}</option>
+          <option value="">{this.props.intl.formatMessage({ id: 'Terra.date.input.monthPlaceholder' })}</option>
           <option key={this.props.intl.formatMessage({ id: 'Terra.date.input.january' })} value="01">{this.props.intl.formatMessage({ id: 'Terra.date.input.january' })}</option>
           <option key={this.props.intl.formatMessage({ id: 'Terra.date.input.february' })} value="02">{this.props.intl.formatMessage({ id: 'Terra.date.input.february' })}</option>
           <option key={this.props.intl.formatMessage({ id: 'Terra.date.input.march' })} value="03">{this.props.intl.formatMessage({ id: 'Terra.date.input.march' })}</option>