+++ packages/terra-date-input/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Fixed
+  * Restricted year input to be between `1900` and `2100`.
+  * Allow users to clear month value by selecting placeholder option value.
+  
 ## 1.22.0 - (October 27, 2020)
 
 * Fixed